get float

+++ js/task-3.js
@@ -1,8 +1,8 @@
 `use strict`
 function getElementWidth(content, padding, border) {
-    const contentNum = parseInt(content);
-    const paddingNum = parseInt(padding);
-    const borderNum = parseInt(border);
+    const contentNum = parseFloat(content);
+    const paddingNum = parseFloat(padding);
+    const borderNum = parseFloat(border);
     return contentNum + paddingNum * 2 + borderNum * 2;
 };
